Added account link to workouts

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,22 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({workout}) => {
   const {dispatch} = useWorkoutsContext();
+  const {user} = useAuthContext();
 
   const handleClick = async () => {
+    // Return if there isn't a user
+    if (!user) {
+      return;
+    }
+
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     });
 
     const json = await response.json();  // json = item deleted
